workspace/quiz/requirements.txt
openpyxl==3.1.2
python-telegram-bot==20.2